+++ CHANGELOG.md
@@ -11,7 +11,7 @@ All notable changes to this project will be documented in this file, following t
 - All color options accept color names and hexcodes
 - `visualStyle` option allows per-component specification
 - Modular UI rendering (5 top-level components renderable separately)
-- Foldseek extension (PDBeMolstarPlugin.extensions.foldseek)
+- Foldseek extension (PDBeMolstarPlugin.extensions.Foldseek)
 - StateGallery extension
 
 ## [v3.2.0] - 2024-04-24